fix: rename staking folder, fix pulumi script ts errors
