+++ infra/lib/amplify/app.ts
@@ -1,11 +1,11 @@
 import * as aws from "@pulumi/aws";
-import { Input } from "@pulumi/pulumi";
+import { Input, Output } from "@pulumi/pulumi";
 import * as pulumi from "@pulumi/pulumi";
 import * as std from "@pulumi/std";
 import * as cloudflare from "@pulumi/cloudflare";
 import * as github from "@pulumi/github";
 
-export function createAmplifyStakingApp(
+export function createAmplifyApp(
   githubUrl: string,
   githubAccessToken: string,
   domainName: string,
@@ -13,9 +13,7 @@ export function createAmplifyStakingApp(
   cloudflareZoneId: string,
   prefix: string,
   branchName: string,
-  environmentVariables: Input<{
-    [key: string]: Input<string>;
-  }>,
+  environmentVariables: Record<string, string | Output<any>>,
   tags: { [key: string]: string },
   enableBasicAuth: boolean,
   username?: string,
@@ -104,12 +102,12 @@ applications:
     const certRecord = domainCert.apply((_cert) => {
       const certDetails = _cert.split(" "); // Name Type Value
       const certRecord = new cloudflare.Record("cloudflare-certificate-record", {
-        name: certDetails[0].replace(`.${cloudflareDomain}.`, ''), // remove the autocomplete domain
+        name: certDetails[0].replace(`.${cloudflareDomain}.`, ""), // remove the autocomplete domain
         zoneId: cloudflareZoneId,
         type: certDetails[1],
         value: certDetails[2],
         allowOverwrite: true,
-        comment: `Certificate for *.${cloudflareDomain}`
+        comment: `Certificate for *.${cloudflareDomain}`,
 
         // ttl: 3600
       });
@@ -125,7 +123,7 @@ applications:
           type: domainDetails[1],
           value: domainDetails[2],
           allowOverwrite: true,
-          comment: `Points to AWS Amplify for stake V2 app`
+          comment: `Points to AWS Amplify for stake V2 app`,
         });
         return record;
       });
